Game engine

+++ final-project/Inc/game_engine.h
@@ -9,13 +9,15 @@
 #define GAME_ENGINE_H_
 /* Private Includes ------------------------- */
 #include "main.h"
+#include <stdio.h>
+#include <stdlib.h>
 
 /* Private enums ------------------------- */
 enum entity { player, enemyUnit };  //0 is player and 1 is enemy
 
 /* Private Structs ------------------------- */
 typedef struct ENTITY {
-	int posit_x;
+	float posit_x;
 	int posit_y;
 	int health;
 	bool enabled;
